// server/src/shared/enums/payment-enum.ts
export enum PaymentMethodType {
  CREDIT_CARD = "CREDIT_CARD",
  DEBIT_CARD = "DEBIT_CARD",
  PAYPAL = "PAYPAL",
  BANK_ACCOUNT = "BANK_ACCOUNT",
  DIGITAL_WALLET = "DIGITAL_WALLET",
  MOMO = "MOMO",
  GIFT_CARD = "GIFT_CARD",
  OTHER = "OTHER",
}

export enum CardBrand {
  VISA = "VISA",
  MASTERCARD = "MASTERCARD",
  AMEX = "AMEX",
  DISCOVER = "DISCOVER",
  JCB = "JCB",
  DINERS_CLUB = "DINERS_CLUB",
  UNIONPAY = "UNIONPAY",
  MAESTRO = "MAESTRO",
  ELO = "ELO",
  OTHER = "OTHER",
}

export enum PaymentStatus {
  PENDING = "PENDING",
  AUTHORIZED = "AUTHORIZED",
  PAID = "PAID",
  PARTIALLY_REFUNDED = "PARTIALLY_REFUNDED",
  REFUNDED = "REFUNDED",
  VOIDED = "VOIDED",
  FAILED = "FAILED",
}